[PM-1200] chore: rename to `vault-timeout`
