+++ apps/browser/src/services/vault-timeout/vault-timeout.service.ts
@@ -1,4 +1,4 @@
-import { VaultTimeoutService as BaseVaultTimeoutService } from "@bitwarden/common/services/vaultTimeout/vaultTimeout.service";
+import { VaultTimeoutService as BaseVaultTimeoutService } from "@bitwarden/common/services/vault-timeout/vault-timeout.service";
 
 import { SafariApp } from "../../browser/safariApp";
