+++ libs/common/src/abstractions/vault-timeout/vault-timeout-settings.service.ts
@@ -1,5 +1,5 @@
 import { VaultTimeoutAction } from "../../enums/vault-timeout-action.enum";
-import { PinLockType } from "../../services/vaultTimeout/vaultTimeoutSettings.service";
+import { PinLockType } from "../../services/vault-timeout/vault-timeout-settings.service";
 
 export abstract class VaultTimeoutSettingsService {
   setVaultTimeoutOptions: (
